支持释放队列
