+++ src/dotnetCampus.FileDownloader/RandomFileWriter.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Buffers;
 using System.IO;
 using System.Runtime.ExceptionServices;
 using System.Threading.Tasks;
@@ -28,7 +27,7 @@ namespace dotnetCampus.FileDownloader
         public void QueueWrite(long fileStartPoint, byte[] data, int dataOffset, int dataLength)
         {
             var fileSegment = new FileSegment(fileStartPoint, data, dataOffset, dataLength);
-            DownloadSegmentList.Enqueue(fileSegment);
+            FileSegmentList.Enqueue(fileSegment);
         }
 
         /// <summary>
@@ -41,7 +40,7 @@ namespace dotnetCampus.FileDownloader
             var task = new TaskCompletionSource<bool>();
 
             var fileSegment = new FileSegment(fileStartPoint, data, 0, data.Length, task);
-            DownloadSegmentList.Enqueue(fileSegment);
+            FileSegmentList.Enqueue(fileSegment);
             await task.Task;
         }
 
@@ -57,7 +56,7 @@ namespace dotnetCampus.FileDownloader
             var task = new TaskCompletionSource<bool>();
 
             var fileSegment = new FileSegment(fileStartPoint, data, dataOffset, dataLength, task);
-            DownloadSegmentList.Enqueue(fileSegment);
+            FileSegmentList.Enqueue(fileSegment);
             await task.Task;
         }
 
@@ -72,7 +71,7 @@ namespace dotnetCampus.FileDownloader
         {
             while (true)
             {
-                var fileSegment = await DownloadSegmentList.DequeueAsync();
+                var fileSegment = await FileSegmentList.DequeueAsync();
 
                 try
                 {
@@ -99,7 +98,7 @@ namespace dotnetCampus.FileDownloader
                     // 执行业务代码
                 }
 
-                var isEmpty = DownloadSegmentList.Count == 0;
+                var isEmpty = FileSegmentList.Count == 0;
 
                 if (isEmpty)
                 {
@@ -116,7 +115,7 @@ namespace dotnetCampus.FileDownloader
 
         private FileStream Stream { get; }
 
-        private AsyncQueue<FileSegment> DownloadSegmentList { get; } = new AsyncQueue<FileSegment>();
+        private AsyncQueue<FileSegment> FileSegmentList { get; } = new AsyncQueue<FileSegment>();
 
         private readonly struct FileSegment
         {
@@ -163,60 +162,56 @@ namespace dotnetCampus.FileDownloader
         /// <returns></returns>
         public async ValueTask DisposeAsync()
         {
-            // 从业务上，这里只有一个线程访问，也不会有重入
-            _isDispose = true;
-
-            if (Exception != null)
+            try
             {
-                ExceptionDispatchInfo.Capture(Exception).Throw();
-            }
-            // 需要先判断 Exception 因为加入只进入一个任务，刚好这个任务炸了
-            if (DownloadSegmentList.Count == 0)
-            {
-                return;
-            }
+                // 从业务上，这里只有一个线程访问，也不会有重入
+                _isDispose = true;
 
-            var task = new TaskCompletionSource<bool>();
-
-            WriteFinished += (sender, args) =>
-            {
                 if (Exception != null)
                 {
-                    task.SetException(Exception);
+                    ExceptionDispatchInfo.Capture(Exception).Throw();
                 }
-                else
+
+                // 需要先判断 Exception 因为加入只进入一个任务，刚好这个任务炸了
+                if (FileSegmentList.Count == 0)
                 {
-                    task.SetResult(true);
+                    return;
                 }
-                WriteFinished = null;
-            };
 
-            if (Exception != null)
-            {
-                throw Exception;
+                var task = new TaskCompletionSource<bool>();
+
+                WriteFinished += (sender, args) =>
+                {
+                    if (Exception != null)
+                    {
+                        task.SetException(Exception);
+                    }
+                    else
+                    {
+                        task.SetResult(true);
+                    }
+
+                    WriteFinished = null;
+                };
+
+                if (Exception != null)
+                {
+                    throw Exception;
+                }
+
+                if (FileSegmentList.Count == 0)
+                {
+                    return;
+                }
+
+                await task.Task;
             }
-
-            if (DownloadSegmentList.Count == 0)
+            finally
             {
-                return;
+                FileSegmentList.Dispose();
             }
-
-            await task.Task;
         }
 
         private bool _isDispose;
-    }
-
-    public static class SharedArrayPool
-    {
-        public static byte[] Rent(int minLength)
-        {
-            return ArrayPool<byte>.Shared.Rent(minLength);
-        }
-
-        public static void Return(byte[] array)
-        {
-            ArrayPool<byte>.Shared.Return(array);
-        }
     }
 }
